работающий blueprism для всех модулей
