+++ webapp/__init__.py
@@ -1,5 +1,7 @@
 from flask import Flask, render_template, flash, redirect, url_for
 from flask_login import LoginManager, current_user, login_required
+from webapp.admin.views import blueprint as admin_blueprint
+from webapp.index.views import blueprint as index_blueprint
 from webapp.user.models import db, User, Person, PythagoreanSquare
 from webapp.user.models import User
 from webapp.user.views import blueprint as user_blueprint
@@ -14,18 +16,8 @@ def create_app():
     login_manager.init_app(app)
     login_manager.login_view = 'user.login'
     app.register_blueprint(user_blueprint)
-
-    @app.route('/admin')
-    @login_required
-    def admin():
-        if current_user.is_admin:
-            return 'Привет админ.'
-        else:
-            return 'Ты не админ'
-
-    @app.route('/index')
-    def index():
-        return 'Здесь строится личный кабинет юзера.'
+    app.register_blueprint(admin_blueprint)
+    app.register_blueprint(index_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
